refactor(conf): put common func to util.go
separate file implementation of conf loader from interface
